pyproject.toml
[project]
name = "vechord"
dynamic = ["version"]
description = "VectorChord Python SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "defspec>=0.3.0",
    "falcon>=4.0.2",
    "httpx>=0.28.1",
    "msgspec>=0.19.0",
    "numpy>=2.0.2",
    "pgvector>=0.3.6",
    "pillow>=11.1.0",
    "psycopg[binary]>=3.2.3",
    "pypdfium2>=4.30.1",
    "pytrec-eval-terrier>=0.5.6",
    "rich>=13.9.4",
]

[project.scripts]
vechord = "vechord.main:main"

[project.optional-dependencies]
gemini = [
    "google-generativeai>=0.8.4",
]
openai = [
    "openai>=1.60.2",
]
spacy = [
    "en-core-web-sm",
    "spacy>=3.8.4",
]
wordllama = [
    "wordllama>=0.3.8.post20",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pdm-backend>=2.4.3",
    "pytest>=8.3.5",
    "ruff>=0.9.1",
]
doc = [
    "furo>=2024.8.6",
    "myst-parser>=4.0.1",
    "sphinx>=8.1.3",
    "sphinx-autodoc-typehints>=3.0.1",
    "sphinx-sitemap>=2.6.0",
    "sphinxcontrib-napoleon>=0.7",
    "sphinxext-opengraph>=0.9.1",
]

[tool.pytest.ini_options]
markers = [
    "db",
]

[tool.ruff]
target-version = "py310"
[tool.ruff.lint]
select = ["E", "F", "G", "B", "I", "SIM", "TID", "PL", "RUF"]
ignore = ["E501"]
[tool.ruff.lint.isort]
known-first-party = ["vechord"]
[tool.ruff.lint.pylint]
max-args = 5
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.10"
warn_redundant_casts = true
warn_unreachable = true
pretty = true

[[tool.mypy.overrides]]
module = [
  "pgvector.psycopg",
  "pytrec_eval",
  "pypdfium2",
  "wordllama",
]
ignore_missing_imports = true

[tool.pdm]
distribution = true
[tool.pdm.version]
source = "scm"
fallback_version = "0.0.0"
write_to = "vechord/__version__.py"
write_template = "__version__ = '{}'"

[tool.typos.default.extend-words]
typ = "typ"

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
